change the package from plans to ast
