+++ ppl/src/main/java/com/amazon/opendistroforelasticsearch/sql/ppl/ast/dsl/DSL.java
@@ -13,25 +13,25 @@
  *   permissions and limitations under the License.
  */
 
-package com.amazon.opendistroforelasticsearch.sql.ppl.plans.dsl;
+package com.amazon.opendistroforelasticsearch.sql.ppl.ast.dsl;
 
-import com.amazon.opendistroforelasticsearch.sql.ppl.plans.expression.AggregateFunction;
-import com.amazon.opendistroforelasticsearch.sql.ppl.plans.expression.And;
-import com.amazon.opendistroforelasticsearch.sql.ppl.plans.expression.DataType;
-import com.amazon.opendistroforelasticsearch.sql.ppl.plans.expression.EqualTo;
-import com.amazon.opendistroforelasticsearch.sql.ppl.plans.expression.Expression;
-import com.amazon.opendistroforelasticsearch.sql.ppl.plans.expression.Function;
-import com.amazon.opendistroforelasticsearch.sql.ppl.plans.expression.In;
-import com.amazon.opendistroforelasticsearch.sql.ppl.plans.expression.Literal;
-import com.amazon.opendistroforelasticsearch.sql.ppl.plans.expression.Map;
-import com.amazon.opendistroforelasticsearch.sql.ppl.plans.expression.Not;
-import com.amazon.opendistroforelasticsearch.sql.ppl.plans.expression.Or;
-import com.amazon.opendistroforelasticsearch.sql.ppl.plans.expression.UnresolvedAttribute;
-import com.amazon.opendistroforelasticsearch.sql.ppl.plans.logical.Aggregation;
-import com.amazon.opendistroforelasticsearch.sql.ppl.plans.logical.Filter;
-import com.amazon.opendistroforelasticsearch.sql.ppl.plans.logical.UnresolvedPlan;
-import com.amazon.opendistroforelasticsearch.sql.ppl.plans.logical.Project;
-import com.amazon.opendistroforelasticsearch.sql.ppl.plans.logical.Relation;
+import com.amazon.opendistroforelasticsearch.sql.ppl.ast.expression.AggregateFunction;
+import com.amazon.opendistroforelasticsearch.sql.ppl.ast.expression.And;
+import com.amazon.opendistroforelasticsearch.sql.ppl.ast.expression.DataType;
+import com.amazon.opendistroforelasticsearch.sql.ppl.ast.expression.EqualTo;
+import com.amazon.opendistroforelasticsearch.sql.ppl.ast.expression.Expression;
+import com.amazon.opendistroforelasticsearch.sql.ppl.ast.expression.Function;
+import com.amazon.opendistroforelasticsearch.sql.ppl.ast.expression.In;
+import com.amazon.opendistroforelasticsearch.sql.ppl.ast.expression.Literal;
+import com.amazon.opendistroforelasticsearch.sql.ppl.ast.expression.Map;
+import com.amazon.opendistroforelasticsearch.sql.ppl.ast.expression.Not;
+import com.amazon.opendistroforelasticsearch.sql.ppl.ast.expression.Or;
+import com.amazon.opendistroforelasticsearch.sql.ppl.ast.expression.UnresolvedAttribute;
+import com.amazon.opendistroforelasticsearch.sql.ppl.ast.tree.Aggregation;
+import com.amazon.opendistroforelasticsearch.sql.ppl.ast.tree.Filter;
+import com.amazon.opendistroforelasticsearch.sql.ppl.ast.tree.UnresolvedPlan;
+import com.amazon.opendistroforelasticsearch.sql.ppl.ast.tree.Project;
+import com.amazon.opendistroforelasticsearch.sql.ppl.ast.tree.Relation;
 import java.util.Arrays;
 import java.util.List;
