+++ ppl/src/main/java/com/amazon/opendistroforelasticsearch/sql/ppl/ast/expression/Map.java
@@ -13,9 +13,9 @@
  *   permissions and limitations under the License.
  */
 
-package com.amazon.opendistroforelasticsearch.sql.ppl.plans.expression;
+package com.amazon.opendistroforelasticsearch.sql.ppl.ast.expression;
 
-import com.amazon.opendistroforelasticsearch.sql.ppl.node.AbstractNodeVisitor;
+import com.amazon.opendistroforelasticsearch.sql.ppl.ast.AbstractNodeVisitor;
 import java.util.Arrays;
 import java.util.List;
 import lombok.EqualsAndHashCode;
@@ -24,25 +24,23 @@ import lombok.RequiredArgsConstructor;
 import lombok.ToString;
 
 /**
- * Expression node of one-to-many mapping relation IN
- * Params include the field expression and/or wildcard field expression, nested field expression (@field)
- * And the values that the field is mapped to (@valueList)
+ * Expression node of one-to-one mapping relation
  */
 @Getter
 @ToString
 @EqualsAndHashCode(callSuper = false)
 @RequiredArgsConstructor
-public class In extends Expression {
-    private final Expression field;
-    private final List<Expression> valueList;
+public class Map extends Expression {
+    private final Expression origin;
+    private final Expression target;
 
     @Override
     public List<Expression> getChild() {
-        return Arrays.asList(field);
+        return Arrays.asList(origin, target);
     }
 
     @Override
     public <R, C> R accept(AbstractNodeVisitor<R, C> nodeVisitor, C context) {
-        return nodeVisitor.visitIn(this, context);
+        return nodeVisitor.visitMap(this, context);
     }
 }
