+++ ppl/src/main/java/com/amazon/opendistroforelasticsearch/sql/ppl/ast/expression/Or.java
@@ -13,23 +13,34 @@
  *   permissions and limitations under the License.
  */
 
-package com.amazon.opendistroforelasticsearch.sql.ppl.plans.logical;
+package com.amazon.opendistroforelasticsearch.sql.ppl.ast.expression;
 
-import com.amazon.opendistroforelasticsearch.sql.ppl.node.AbstractNodeVisitor;
-import com.amazon.opendistroforelasticsearch.sql.ppl.node.Node;
+import com.amazon.opendistroforelasticsearch.sql.ppl.ast.AbstractNodeVisitor;
+import java.util.Arrays;
+import java.util.List;
 import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 import lombok.ToString;
 
 /**
- * Abstract unresolved plan
+ * Expression node of the logic OR
  */
-@EqualsAndHashCode(callSuper = false)
+@Getter
 @ToString
-public abstract class UnresolvedPlan extends Node {
+@EqualsAndHashCode(callSuper = false)
+@RequiredArgsConstructor
+public class Or extends Expression {
+    private final Expression left;
+    private final Expression right;
+
     @Override
-    public <T, C> T accept(AbstractNodeVisitor<T, C> nodeVisitor, C context) {
-        return nodeVisitor.visitChildren(this, context);
+    public List<Expression> getChild() {
+        return Arrays.asList(left, right);
     }
 
-    public abstract UnresolvedPlan attach(UnresolvedPlan child);
+    @Override
+    public <R, C> R accept(AbstractNodeVisitor<R, C> nodeVisitor, C context) {
+        return nodeVisitor.visitOr(this, context);
+    }
 }
