+++ ppl/src/main/java/com/amazon/opendistroforelasticsearch/sql/ppl/ast/tree/Relation.java
@@ -13,18 +13,38 @@
  *   permissions and limitations under the License.
  */
 
-package com.amazon.opendistroforelasticsearch.sql.ppl.plans.expression;
+package com.amazon.opendistroforelasticsearch.sql.ppl.ast.tree;
 
-import com.amazon.opendistroforelasticsearch.sql.ppl.node.AbstractNodeVisitor;
-import com.amazon.opendistroforelasticsearch.sql.ppl.node.Node;
+import com.amazon.opendistroforelasticsearch.sql.ppl.ast.AbstractNodeVisitor;
+import com.google.common.collect.ImmutableList;
+import java.util.List;
 import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 import lombok.ToString;
 
-@EqualsAndHashCode(callSuper = false)
+/**
+ * Logical plan node of Relation, the interface for building the searching sources
+ */
+@Getter
 @ToString
-public abstract class Expression extends Node {
+@EqualsAndHashCode(callSuper = false)
+@RequiredArgsConstructor
+public class Relation extends UnresolvedPlan {
+    private final String tableName;
+
+    @Override
+    public List<UnresolvedPlan> getChild() {
+        return ImmutableList.of();
+    }
+
     @Override
     public <T, C> T accept(AbstractNodeVisitor<T, C> nodeVisitor, C context) {
-        return nodeVisitor.visitChildren(this, context);
+        return nodeVisitor.visitRelation(this, context);
+    }
+
+    @Override
+    public UnresolvedPlan attach(UnresolvedPlan child) {
+        return this;
     }
 }
